Add Regal for linting Rego
Hey there brainiacs! 👋😃 This PR introduces [Regal](https://github.com/styrainc/regal)
for linting the Rego included in this project. A lot of Rego to process here, and quite a
few reported violations. I've added a Regal configuration file to ignore a bunch of rules
for the time being, and fixed some of the more low-hanging fruits as part of the PR.

The rules addressed in this PR:

* [use-some-for-output-vars](https://docs.styra.com/regal/rules/idiomatic/use-some-for-output-vars)
* [no-whitespace-comment](https://docs.styra.com/regal/rules/style/no-whitespace-comment)
* [non-raw-regex-pattern](https://docs.styra.com/regal/rules/idiomatic/non-raw-regex-pattern)
* [unconditional-assignment](https://docs.styra.com/regal/rules/style/unconditional-assignment)
* [equals-pattern-matching](https://docs.styra.com/regal/rules/idiomatic/equals-pattern-matching)
* [identically-named-tests](https://docs.styra.com/regal/rules/testing/identically-named-tests)

See the config file provided for a list of the rules ignored, where I've tried to annotate
the rules with some short explanations, and pointers to the docs.

I've also added steps to lint Rego as part of pushes/PRs.

Let me know what you think!

Signed-off-by: Anders Eknert <anders@styra.com>